requirements.txt
Django==4.1.*
djangorestframework==3.14.*
celery==5.2.*
redis==4.3.*
django-summernote==0.8.*
psycopg2-binary==2.9.*
django-celery-beat==2.7.0
django-filter==23.1
gunicorn==20.1.0
requests==2.32.3